Use getOrigin to use proper headers and move allowedOrigin to existing origin lib
